New type of queue, more like the order book

+++ Bets/Ticket.java
@@ -3,12 +3,12 @@ import java.util.Date;
 public class Ticket{
 	private String match;
 	private int outcome;
-	private double prize;
+	private int prize;
 	private Date date;
 	private int amount;
 	private boolean locked;
 
-	public Ticket(String match, int outcome, double prize, int amount, Date date)
+	public Ticket(String match, int outcome, int prize, int amount, Date date)
 	{
 		this.match = match;
 		this.prize = prize;
@@ -24,7 +24,7 @@ public class Ticket{
 		this.outcome = outcome;
 	}
 
-	public Ticket(String match, int outcome, double prize, int amount)
+	public Ticket(String match, int outcome, int prize, int amount)
 	{
 		this.match = match;
 		this.prize = prize;
@@ -64,7 +64,7 @@ public class Ticket{
 		return this.date;
 	}
 	
-	public double getPrize()
+	public int getPrize()
 	{
 		return this.prize;
 	}
